requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
asyncpg==0.30.0
pydantic-settings==2.8.1
boto3==1.38.3
python-multipart==0.0.20
email-validator==2.2.0
